requirements.txt
torch
torchaudio

nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git@main

numpy<2

sphn

plotly
